# docs/swagger.yaml
definitions:
  internal_api.GetProductResponse:
    properties:
      attributes:
        $ref: '#/definitions/pgtype.Hstore'
      categoryId:
        type: integer
      categoryImage:
        type: string
      categoryName:
        type: string
      description:
        type: string
      discount:
        type: integer
      images:
        items:
          type: string
        type: array
      nameEn:
        type: string
      nameFa:
        type: string
      price:
        type: integer
      priceOff:
        type: integer
      quantity:
        type: integer
    type: object
  internal_api.HTTPError:
    properties:
      message:
        type: string
    type: object
  internal_api.ListCategoriesRow:
    properties:
      image:
        type: string
      nameEn:
        type: string
      nameFa:
        type: string
    type: object
  internal_api.ListProductsRow:
    properties:
      discount:
        type: integer
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      outOfStock:
        type: boolean
      price:
        type: integer
      priceOff:
        type: integer
    type: object
  pgtype.Hstore:
    additionalProperties:
      type: string
    type: object
info:
  contact: {}
paths:
  /public/categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api.ListCategoriesRow'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.HTTPError'
      summary: Lists the categories
      tags:
      - categories
  /public/products:
    get:
      parameters:
      - description: Product list order
        enum:
        - name
        - price_highest
        - price_lowest
        in: query
        name: orderBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api.ListProductsRow'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.HTTPError'
      summary: Lists the products
      tags:
      - products
  /public/products/{id}:
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.GetProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.HTTPError'
      summary: Fetches a product info
      tags:
      - products
swagger: "2.0"