# internal/api/database/categories.sql.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0
// source: categories.sql

package database

import (
	"context"
)

const listCategories = `-- name: ListCategories :many
select
    image,
    name_fa,
    name_en
from
    categories
`

type ListCategoriesRow struct {
	Image  string `json:"image"`
	NameFa string `json:"nameFa"`
	NameEn string `json:"nameEn"`
}

func (q *Queries) ListCategories(ctx context.Context) ([]ListCategoriesRow, error) {
	rows, err := q.db.Query(ctx, listCategories)
	if err != nil {
		return nil, err
	}
	defer rows.Close()
	items := []ListCategoriesRow{}
	for rows.Next() {
		var i ListCategoriesRow
		if err := rows.Scan(&i.Image, &i.NameFa, &i.NameEn); err != nil {
			return nil, err
		}
		items = append(items, i)
	}
	if err := rows.Err(); err != nil {
		return nil, err
	}
	return items, nil
}